+++ lib/rorvswild/plugin/active_record.rb
@@ -11,7 +11,7 @@ module RorVsWild
 
       def self.setup_callback
         return unless defined?(::ActiveSupport::Notifications.subscribe)
-        ActiveSupport::Notifications.subscribe("sql.active_record", new)
+        ::ActiveSupport::Notifications.subscribe("sql.active_record", new)
       end
 
       IGNORED_QUERIES = %w[EXPLAIN SCHEMA].freeze